PP-6627 rename valid_account_id to permissions
- rename valid_account_id to a more meaningful name

with @sfount
