+++ app/controllers/all-service-transactions/get-controller.js
@@ -7,7 +7,7 @@ const { response, renderErrorView } = require('../../utils/response')
 const { ConnectorClient } = require('../../services/clients/connector_client.js')
 const transactionService = require('../../services/transaction_service')
 const { buildPaymentList } = require('../../utils/transaction_view.js')
-const { liveUserServicesGatewayAccounts } = require('./../../utils/valid_account_id')
+const { liveUserServicesGatewayAccounts } = require('../../utils/permissions')
 const { getFilters, describeFilters } = require('../../utils/filters.js')
 const router = require('../../routes.js')
 const states = require('../../utils/states')
